standardize voltage units to mV

+++ sensor_software/src/adc_handler.h
@@ -21,6 +21,8 @@
 #ifndef ADC_HANDLER_H
 #define ADC_HANDLER_H
 
+#include "common_defines.h"
+
 #include "stdint.h"
 
 // Halcogen generated SPI driver
@@ -68,12 +70,12 @@ void adc_set_control_reg(ADC_Handler *handl, uint8_t repeat,
 void adc_get_raw(ADC_Handler *handl, uint16_t *data, uint8_t *ch);
 
 
-float adc_conv_to_volt(uint16_t value, float vref);
-float adc_conv_to_celsius(uint16_t value, float vref);
-float adc_conv_to_current(uint16_t value, float vref);
+int16_t adc_conv_to_volt(uint16_t value, int16_t vref);
+float adc_conv_to_celsius(uint16_t value, int16_t vref);
+int16_t adc_conv_to_current(uint16_t value, int16_t vref);
 
 // convert internal temp sensor value
-float adc_get_tsense_temp(uint16_t value, float vref);
+float adc_get_tsense_temp(uint16_t value, int16_t vref);
 
 // SPI Read/Write functions
 void write_spi(ADC_Handler *handl, uint16_t *data, uint32_t len);
